Fixes inverted colors (Red/Blue swap) after switching to metal on macOS
The Kitty Graphics Protocol strictly requires RGBA pixel data, but the Flutter engine (via Skia) on macOS is hardcoded to output BGRA textures (kBGRA_8888_SkColorType). Changing the requested texture format to RGBA causes Skia surface creation to fail.

Previously, we were swapping color channels on the CPU (in terminal_window.rs in `draw_kitty` - broken code which was enabled only on Linux but was actually needed on macOS), which required iterating over every pixel of every frame. Reenabling it on macOS to fix the correctness issue is extremely slow and caused high CPU usage. It takes 20+ms to draw a single frame now in the embedder.

This commit implements a Metal Compute Shader (`copy_texture_to_buffer`) to handle the conversion on the GPU:
- The shader reads the BGRA texture and writes RGBA bytes to a shared
  `MTLBuffer`.
- CPU access is optimized using a persistent  with appropriate synchronization (handling `MTLStorageModeManaged` for Intel Macs).
- The slow CPU-side color swapping loop has been removed.

This results in correct colors and significantly improved rendering performance.